fix: no more theme flickering while loading, enabling transitions 100ms after loading fix: invited users list now looks ok fix: rating overlay can grow vertically up to 320px if needed refactor: theme generation functionality moved to theme provider refactor: removed unused js-transitions code
