+++ frontend/src/Components/PostComponent.module.scss
@@ -44,10 +44,9 @@
     }
     .username {
         color: var(--fgMedium);
-        svg{ fill: var(--fgSoftest); }
+        svg{ fill: var(--fgMedium); }
         &:hover {
             color: var(--primary);
-            svg{  fill: var(--primary); }
         }
     }
 }
