+++ frontend/src/Pages/UserPage.module.scss
@@ -7,6 +7,7 @@
 .header, .userinfo {
     padding: 10px 0;
     >div{
+        width: 100%;
         display: flex;
         flex-wrap: wrap;
         gap: 8px;
