+++ frontend/src/theme.ts
@@ -1,6 +1,4 @@
-import {ThemeCollection, ThemeStyles} from './Theme/ThemeProvider';
-import rgba from 'color-normalize';
-import colorspace from 'color-space';
+import {ThemeCollection} from './Theme/ThemeProvider';
 
 /**
 
@@ -22,7 +20,7 @@ preview of all theme colors is available at http://orbitar.local/theme
 
 */
 
-const themes: ThemeCollection = {
+export const themes: ThemeCollection = {
     debugTheme: {
         colors: {
 
@@ -349,101 +347,4 @@ const themes: ThemeCollection = {
             }
         },
     },
-};
-
-
-export const getThemes = () => {
-    Object.values(themes).map( preprocessTheme );
-    return themes;
-};
-
-// GENERATE THEME
-const preprocessTheme = (theme: ThemeStyles) => {
-    const colors = theme.colors;
-
-    //detect if this is dark theme or light theme
-    const fg = colors.fg as string;
-    const isDark: boolean = hsl(fg).l > 0.5;
-
-    // generate harder versions of foreground color
-    colors.fgHardest ??= isDark ? '#fff':'#000';
-    colors.fgHard ??= lerpColor(fg, colors.fgHardest as string, .33 );
-    colors.fgHarder ??= lerpColor(fg, colors.fgHardest as string, .66 );
-
-    // generate softer versions of foreground color
-    colors.fgMedium ??= reduceAlpha( fg, .2);  //-10% alpha
-    colors.fgSoft ??= reduceAlpha( fg, .3);
-    colors.fgSofter ??= reduceAlpha( fg, .4);
-    colors.fgSoftest ??= reduceAlpha( fg, .5);
-    colors.fgGhost ??= reduceAlpha( fg, .7);
-    colors.fgAlmostInvisible ??= reduceAlpha( fg, .9); //-90% alpha
-
-    colors.onAccent ??= '#fff';
-    colors.onAccentGhost ??= 'rgba(255,255,255,0.7)';
-
-    // generate primary variants
-    colors.primaryHover ??= increaseSaturation( colors.primary as string, 0.75 );  //+75% saturation
-    colors.primaryGhost ??= reduceAlpha( colors.primary as string, .7); //-70% alpha
-
-    // generate danger variants
-    colors.dangerHover ??= increaseSaturation( colors.danger as string, 0.75 );
-    colors.dangerGhost ??= reduceAlpha( colors.danger as string, .7);
-
-    // generate positive variants
-    colors.positiveHover ??= increaseSaturation( colors.positive as string, 0.75 );
-    colors.positiveGhost ??= reduceAlpha( colors.positive as string, .7);
-
-    // generate link variants
-    colors.linkHover ??= increaseSaturation( colors.link as string, 0.75 );
-    colors.linkGhost ??= reduceAlpha( colors.link as string, .7);
-
-    // backgrounds
-    const c = isDark ? 255 : 0;
-    colors.elevated ??= isDark ? lerpColor( colors.bg as string, '#ffffff', 0.03): '#fff';
-    colors.lowered ??= lerpColor( colors.bg as string, '#000000',  isDark ? 0.2 : 0.03 );
-    colors.dim1 ??= rgbaToString([c,c,c,0.02]);
-    colors.dim2 ??= rgbaToString([c,c,c,0.04]);
-    colors.dim3 ??= rgbaToString([c,c,c,0.06]);
-};
-
-const reduceAlpha = (c: string, r: number): string => {
-    const [red,g,b,a] = rgba(c);
-    return rgbaToString( [red,g,b, lerp(a,0,r)]);
-};
-
-const increaseSaturation = (c: string, r: number): string => {
-    const p = hsl(c);
-    p.s = lerp(p.s, 1, r);
-    return HSLtoRGBString(p);
-};
-
-const lerp = (a: number, b: number, r: number) =>  a + Math.min(Math.max(r,0),1) * (b-a);
-
-const lerpColor = (a: string, b: string, r: number) => {
-    const c1 = rgba( a );
-    const c2 = rgba( b );
-    const bal = Math.min(Math.max(r,0),1);
-    const res = [
-        c1[0] + bal * (c2[0]-c1[0]),
-        c1[1] + bal * (c2[1]-c1[1]),
-        c1[2] + bal * (c2[2]-c1[2]),
-        c1[3] + bal * (c2[3]-c1[3])
-    ];
-    return rgbaToString(res);
-};
-
-const rgbaToString = (color: number[]): string => {
-    return `rgba(${Math.floor(color[0] * 255)}, ${Math.floor(color[1] * 255)}, ${Math.floor(color[2] * 255)}, ${color[3].toFixed(2).replace(/\.?0+$/, '')})`;
-};
-
-type HSL = { h: number, s: number, l: number };
-const hsl = (color: string): HSL => {
-    const [r,g,b] = rgba(color);
-    const hsl = colorspace.rgb.hsl([r * 255, g * 255, b * 255]);
-    return { h: hsl[0] / 360, s: hsl[1] / 100, l: hsl[2] / 100 };
-};
-
-const HSLtoRGBString = (c: HSL): string => {
-    const rgb = colorspace.hsl.rgb([c.h * 360, c.s * 100, c.l * 100]);
-    return `rgb(${rgb[0]}, ${rgb[1]}, ${rgb[2]})`;
 };
